+++ packages/web/src/pages/search-page-v2/SearchHeader.tsx
@@ -1,8 +1,7 @@
-import { ChangeEvent, ReactElement, useCallback, useState } from 'react'
+import { ChangeEvent, useCallback } from 'react'
 
-import { GENRES, Maybe, convertGenreLabelToValue } from '@audius/common/utils'
+import { Maybe } from '@audius/common/utils'
 import {
-  OptionsFilterButton,
   Flex,
   IconAlbum,
   IconNote,
@@ -10,21 +9,15 @@ import {
   IconUser,
   RadioGroup,
   SelectablePill,
-  Text,
-  FilterButton,
-  Popup,
-  Paper,
-  TextInput,
-  SegmentedControl,
-  IconCaretDown
+  Text
 } from '@audius/harmony'
 import { capitalize } from 'lodash'
-import { useSearchParams } from 'react-router-dom-v5-compat'
 
 import Header from 'components/header/desktop/Header'
 import { useMedia } from 'hooks/useMedia'
 
-import { Category, Filter } from './types'
+import { Category } from './types'
+import { filters } from './SearchFilters'
 
 export const categories = {
   all: { filters: [] },
@@ -38,211 +31,6 @@ export const categories = {
 } satisfies Record<string, Category>
 
 export type CategoryKey = keyof typeof categories
-
-const messages = {
-  genre: 'Genre',
-  mood: 'Mood',
-  key: 'Key',
-  bpm: 'BPM',
-  isPremium: 'Premium',
-  isVerified: 'Verified',
-  hasDownloads: 'Downloads Available'
-}
-
-const urlSearchParamsToObject = (
-  searchParams: URLSearchParams
-): Record<string, string> =>
-  [...searchParams.entries()].reduce(
-    (result, [key, value]) => ({
-      ...result,
-      [key]: value
-    }),
-    {}
-  )
-
-const useUpdateSearchParams = (key: string) => {
-  const [searchParams, setUrlSearchParams] = useSearchParams()
-  return (value: string) => {
-    if (value) {
-      // TODO: This is causing an amplitude page view every time
-      // let's fix this
-      setUrlSearchParams({
-        ...urlSearchParamsToObject(searchParams),
-        [key]: value
-      })
-    } else {
-      const { [key]: ignored, ...params } =
-        urlSearchParamsToObject(searchParams)
-      setUrlSearchParams(params)
-    }
-  }
-}
-
-const GenreFilter = () => {
-  const [urlSearchParams] = useSearchParams()
-  const genre = urlSearchParams.get('genre')
-  const updateSearchParams = useUpdateSearchParams('genre')
-
-  return (
-    <OptionsFilterButton
-      label={messages.genre}
-      popupAnchorOrigin={{ vertical: 'bottom', horizontal: 'left' }}
-      popupMaxHeight={400}
-      popupTransformOrigin={{ vertical: 'top', horizontal: 'left' }}
-      selection={genre}
-      onChange={updateSearchParams}
-      options={GENRES.map((genre) => ({
-        label: genre,
-        value: convertGenreLabelToValue(genre)
-      }))}
-      showFilterInput
-      filterInputPlaceholder='Search genre'
-    />
-  )
-}
-
-const BpmFilter = () => {
-  const [urlSearchParams] = useSearchParams()
-  const bpm = urlSearchParams.get('bpm')
-  const updateSearchParams = useUpdateSearchParams('bpm')
-
-  const [bpmFilterType, setBpmFilterType] = useState<'range' | 'target'>(
-    'range'
-  )
-
-  const label = bpm ? `${bpm} ${messages.bpm}` : `${messages.bpm}`
-
-  const handleBpmInputChange = useCallback(
-    (handleChange: (value: string, label: string) => void) =>
-      (value: string) => {
-        handleChange(value, `${value} ${messages.bpm}`)
-      },
-    []
-  )
-
-  return (
-    <FilterButton
-      value={bpm}
-      label={label}
-      onChange={updateSearchParams}
-      iconRight={IconCaretDown}
-    >
-      {({ handleChange, isOpen, setIsOpen, anchorRef }) => (
-        <Popup
-          anchorRef={anchorRef}
-          isVisible={isOpen}
-          onClose={() => setIsOpen(false)}
-          anchorOrigin={{ vertical: 'bottom', horizontal: 'left' }}
-          transformOrigin={{ vertical: 'top', horizontal: 'left' }}
-        >
-          <Paper mt='s' border='strong' shadow='far'>
-            <Flex
-              p='s'
-              direction='column'
-              alignItems='flex-start'
-              role='listbox'
-            >
-              <Flex direction='column' w='100%' gap='s'>
-                <SegmentedControl
-                  options={[
-                    { key: 'range', text: 'Range' },
-                    { key: 'target', text: 'Target' }
-                  ]}
-                  selected={bpmFilterType}
-                  onSelectOption={setBpmFilterType}
-                />
-                <TextInput
-                  placeholder={messages.bpm}
-                  label={messages.bpm}
-                  onClick={(e) => {
-                    e.stopPropagation()
-                  }}
-                  onChange={(e) => {
-                    handleBpmInputChange(handleChange)(e.target.value)
-                  }}
-                />
-              </Flex>
-            </Flex>
-          </Paper>
-        </Popup>
-      )}
-    </FilterButton>
-  )
-}
-
-const IsPremiumFilter = () => {
-  const [urlSearchParams] = useSearchParams()
-  const isPremium = urlSearchParams.get('isPremium')
-  const updateSearchParams = useUpdateSearchParams('isPremium')
-
-  return (
-    <FilterButton
-      label={messages.isPremium}
-      value={isPremium}
-      onClick={() => updateSearchParams(isPremium ? '' : 'true')}
-    ></FilterButton>
-  )
-}
-
-const HasDownloadsFilter = () => {
-  const [urlSearchParams] = useSearchParams()
-  const hasDownloads = urlSearchParams.get('hasDownloads')
-  const updateSearchParams = useUpdateSearchParams('hasDownloads')
-
-  return (
-    <FilterButton
-      label={messages.hasDownloads}
-      value={hasDownloads}
-      onClick={() => {
-        updateSearchParams(hasDownloads ? '' : 'true')
-      }}
-    ></FilterButton>
-  )
-}
-
-const IsVerifiedFilter = () => {
-  const [urlSearchParams] = useSearchParams()
-  const isVerified = urlSearchParams.get('isVerified')
-  const updateSearchParams = useUpdateSearchParams('isVerified')
-
-  return (
-    <FilterButton
-      label={messages.isVerified}
-      value={isVerified}
-      onClick={() => {
-        updateSearchParams(isVerified ? '' : 'true')
-      }}
-    ></FilterButton>
-  )
-}
-
-const filters: Record<Filter, () => ReactElement> = {
-  genre: GenreFilter,
-  mood: () => (
-    <OptionsFilterButton
-      label={messages.mood}
-      options={[
-        {
-          value: 'Filter'
-        }
-      ]}
-    />
-  ),
-  key: () => (
-    <OptionsFilterButton
-      label={messages.key}
-      options={[
-        {
-          value: 'Filter'
-        }
-      ]}
-    />
-  ),
-  bpm: BpmFilter,
-  isPremium: IsPremiumFilter,
-  hasDownloads: HasDownloadsFilter,
-  isVerified: IsVerifiedFilter
-}
 
 type SearchHeaderProps = {
   category?: CategoryKey
